SmartNutri Backend/requirements.txt
asgiref==3.8.1
cramjam==2.9.0
Django==5.0.4
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.3
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
mongoengine==0.29.1
pillow==10.3.0
PyJWT==2.9.0
pymongo==4.10.1
PyMySQL==1.1.1
python-snappy==0.7.3
sqlparse==0.5.1
typing_extensions==4.11.0
tzdata==2024.1
winkerberos==0.12.0

